# services/user-service/docs/swagger.yaml
basePath: /
definitions:
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.BulkUserImportRequest:
    properties:
      users:
        items:
          $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.CreateUserRequest'
        minItems: 1
        type: array
    required:
    - users
    type: object
...
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.CreateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
      register_no:
        type: integer
      role_id:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - register_no
    - role_id
    type: object
...
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.ProfileResponse:
    properties:
      bio:
        type: string
      created_at:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone:
        type: string
      profile_id:
        type: string
      profile_picture_url:
        type: string
      register_no:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.UpdateProfileRequest:
    properties:
      bio:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone:
        type: string
      profile_picture_url:
        type: string
    type: object
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.UpdateUserRequest:
    properties:
      email:
        type: string
      role_id:
        type: string
      status:
        enum:
        - active
        - inactive
        - suspended
        type: string
    type: object
  github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.UserWithProfileResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      last_login:
        type: string
      profile:
        $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.ProfileResponse'
      register_no:
        type: integer
      role_id:
        type: string
      role_name:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/utils.Pagination'
      success:
        type: boolean
    type: object
  utils.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8081
info:
  contact:
    email: support@nimbusu.edu
    name: API Support
    url: http://www.nimbusu.edu/support
  description: User Management Service for NimbusU University Platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: NimbusU User Service API
  version: "1.0"
paths:
  /admin/users:
    get:
      consumes:
      - application/json
      description: Get a paginated list of users with filtering (Admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by Role ID
        in: query
        name: role_id
        type: string
      - description: Filter by Status
        in: query
        name: status
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.UserWithProfileResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List Users
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Create a new user (Admin only)
      parameters:
      - description: User Creation Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create User
      tags:
      - admin
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user by ID (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - admin
    get:
      consumes:
      - application/json
      description: Get user details by ID (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.UserWithProfileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get User by ID
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update user details by ID (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - admin
  /admin/users/{id}/activate:
    post:
      consumes:
      - application/json
      description: Activate a suspended or inactive user (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Activate User
      tags:
      - admin
  /admin/users/{id}/suspend:
    post:
      consumes:
      - application/json
      description: Suspend an active user (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Suspend User
      tags:
      - admin
  /admin/users/bulk-import:
    post:
      consumes:
      - application/json
      description: Import multiple users at once (Admin only)
      parameters:
      - description: Bulk Import Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.BulkUserImportRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Bulk Import Users
      tags:
      - admin
...
  /users/me:
    get:
      consumes:
      - application/json
      description: Get the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.UserWithProfileResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get My Profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the profile of the currently authenticated user
      parameters:
      - description: Profile Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_SureshAmal_NimbusU-backend_services_user-service_internal_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update My Profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"